fix builder default config
